+++ src/common/components/images/DefaultWorkspaceProfileImage/DefaultWorkspaceProfileImage.client.tsx
@@ -4,7 +4,7 @@ import clsx from 'clsx';
 
 import { PROFILE_COLORS } from '@common/constants/profile.constants';
 
-import hashCode from '@common/utils/hash-code';
+import hashCode from '@common/utils/hash-code.utils';
 
 import { DefaultWorkspaceProfileImageProps } from './DefaultWorkspaceProfileImage.types';
