+++ src/common/components/inputs/input-invite-member/InputInviteMember/InputInviteMember.client.tsx
@@ -4,11 +4,11 @@ import { useState } from 'react';
 
 import clsx from 'clsx';
 
+import { isValidEmail } from '@common/utils/valid-email.utils';
+
 import InviteButton from '@common/components/buttons/32px/InviteButton/InviteButton.client';
 import EmailChip from '@common/components/inputs/input-invite-member/emails/EmailChip/EmailChip.client';
 import EmailTag from '@common/components/inputs/input-invite-member/emails/EmailTag/EmailTag.client';
-
-import { isValidEmail } from '@common/utils/valid-email-utils';
 
 import { InputInviteMemberProps } from './InputInviteMember.types';
