+++ src/common/utils/hash-code.utils.ts
